ref: #4 refactor
