+++ presentation/view/businesscard.vcf.sample
